requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
yfinance
pandas
pyarrow
minio
mlflow==2.9.2
scikit-learn